requirements.txt
numpy>=1.20.0
matplotlib>=3.4.0
opencv-python>=4.5.0
scipy>=1.7.0
pygame>=2.0.0
pyyaml>=6.0
